review fixes: rework time utils
